cleanup: split Fortia.js into Session.js and SessionManager.js; delete dead code
